Refactor analytical modal solver
Refactor analytical modal solver
